+++ packages/web3-eth-ens/test/integration/resolver.test.ts
@@ -37,8 +37,8 @@ import {
 	itIf,
 } from '../fixtures/system_tests_utils';
 
-import { ENSRegistryAbi } from '../../src/abi/ens/ENSRegistry';
-import { PublicResolverAbi } from '../../src/abi/ens/PublicResolver';
+import { ENSRegistryAbi } from '../fixtures/ens/abi/ENSRegistry';
+import { PublicResolverAbi } from '../fixtures/ens/abi/PublicResolver';
 import { NameWrapperAbi } from '../fixtures/ens/abi/NameWrapper';
 import { ENSRegistryBytecode } from '../fixtures/ens/bytecode/ENSRegistryBytecode';
 import { NameWrapperBytecode } from '../fixtures/ens/bytecode/NameWrapperBytecode';
@@ -169,12 +169,9 @@ describe('ens', () => {
 	});
 
 	it('fetch pubkey', async () => {
-		await ens.setResolver(
-			domain,
-			resolver.options.address as string,
-			sendOptions,
-			DEFAULT_RETURN_FORMAT,
-		);
+		await registry.methods
+			.setResolver(domain, resolver.options.address as string)
+			.send(sendOptions);
 
 		const res = await ens.getPubkey(domain);
 
@@ -186,7 +183,7 @@ describe('ens', () => {
 		const x = '0x1000000000000000000000000000000000000000000000000000000000000000';
 		const y = '0x2000000000000000000000000000000000000000000000000000000000000000';
 
-		await ens.setPubkey(domain, x, y, sendOptions);
+		await resolver.methods.setPubkey(domain, x, y).send(sendOptions);
 
 		const result = await ens.getPubkey(domain);
 
@@ -195,7 +192,7 @@ describe('ens', () => {
 	});
 
 	it('sets contenthash', async () => {
-		await ens.setContenthash(domain, contentHash, sendOptions);
+		await resolver.methods.setContenthash(domain, contentHash).send(sendOptions);
 
 		const res = await resolver.methods.contenthash(domainNode).call(sendOptions);
 		expect(res).toBe(contentHash);
@@ -211,7 +208,7 @@ describe('ens', () => {
 			event.on('data', () => {
 				resolve();
 			});
-			await ens.setContenthash(domain, contentHash, sendOptions);
+			await resolver.methods.setContenthash(domain, contentHash).send(sendOptions);
 		});
 	});
 
@@ -227,7 +224,7 @@ describe('ens', () => {
 			.setResolver(domainNode, resolver.options.address as string)
 			.send(sendOptions);
 
-		await ens.setAddress(domain, accounts[1], sendOptions, DEFAULT_RETURN_FORMAT);
+		await resolver.methods.setAddr(domain, accounts[1]).send(sendOptions);
 
 		const res = await resolver.methods.addr(domainNode, DEFAULT_COIN_TYPE).call(sendOptions);
 		expect(res).toBe(accounts[1]);