+++ CHANGELOG.md
@@ -1465,6 +1465,7 @@ should use 4.0.1-alpha.0 for testing.
 ### Breaking Changes
 
 -   dropped support for NodeJs@14
+-   removed non read-only methods from ens package
 
 ### Added
 
@@ -1488,3 +1489,9 @@ should use 4.0.1-alpha.0 for testing.
 
 -   `formatTransaction` no longer throws a `TransactionDataAndInputError` if it's passed a transaction object with both `data` and `input` properties set (as long as they are the same value) (#6064)
 -   Refactored documentation for `rpc_method_wrappers` to point to the previously duplicated documentation found under the `Web3Eth` class documentation (#6054)
+
+#### web3-eth-ens
+
+### Changed
+
+-   Removed non read-only methods